+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add `listenAsync` for promised listen
+
+### Changed
+- **[BREAKING]** remove `listen` patch
 
 ## [0.3.0] - 2019-10-31
 ### Fixed